move GameContext to folder
